requirements.txt
playwright
python-dotenv
pyairtable
python-telegram-bot
flask
